Minor tweaks to data types and some usability enhancements

+++ gaiku-3d/src/bakers/heightmap.rs
@@ -1,6 +1,6 @@
 use std::collections::HashMap;
 
-use crate::common::{mint::Vector3, Baker, Chunk, Mesh};
+use crate::common::{Baker, Chunk, Mesh};
 
 pub struct HeightMapBaker;
 
@@ -40,7 +40,7 @@ impl Baker for HeightMapBaker {
 
         let mut vertices = vec![[0.0, 0.0, 0.0].into(); vertices_cache.len()];
         for (_, (vertex, index)) in vertices_cache {
-            vertices[index] = vertex.clone();
+            vertices[index as usize] = vertex.clone();
         }
 
         if indices.len() > 0 {
