+++ gaiku-3d/src/bakers/voxel/mod.rs
@@ -125,7 +125,7 @@ impl Baker for VoxelBaker {
 
         let mut vertices: Vec<Vector3<f32>> = vec![[0.0, 0.0, 0.0].into(); vertices_cache.len()];
         for (_, (vertex, index)) in vertices_cache {
-            vertices[index] = vertex.clone();
+            vertices[index as usize] = vertex.clone();
         }
 
         if indices.len() > 0 {
